+++ Idle Boss Fight/Assets/Scripts/Player/PlayerData.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 public class PlayerData : MonoBehaviour
 {
     [field: SerializeField]
-    public int AttackPower { get; set; } = 1;
+    public float AttackPower { get; set; } = 1;
     [field: SerializeField]
-    public int ClickPower { get; set; } = 1;
+    public float ClickPower { get; set; } = 1;
     [field: SerializeField]
-    public float AttackRate { get; set; } = 1f;
+    public float AttackRate { get; set; } = 1;
 
 }
